Implement reverse iterator
For our use-case, we always want to traverse the set in descending sorted order. Adding a getReverseIntIterator eases this operation.

I've also added guava as a test-only dependency to save some code; if this is a problem I'm happy to restore the earlier code + implement some form of Lists.reverse.