jmh/run.sh: make mvn clean optional
This commit makes cleaning of JARs used for the benchmarks optional.
In most of cases Maven should do a pretty good job of building a new JAR
without the need for cleaning the target directory.

This cuts the time it takes to start benchmarks from ~40s to ~30s
on my computer