Roaring64Bitmap::addInt Setting Wrong Bits (#448)
Since int is smaller than long, an int should always map to a positive
number in a long, that is, converting an int to a long should produce
a long in the range of [0, 2^32]. Since addInt directly calls addLong
without doing a proper conversion, the negative sign is being carried
over and setting incorrect positions in the map. For example,
addInt(-1) should be equivilant to addLong(((int) -1) & 0xffffffffL),
but its not. This patch fixes the issue by converting an int to
an unsigned long before calling addLong.