Roaring64Bitmap Memory Leak
Since Roaring64Bitmap only adds to the ART and never removes a
prefix even if the lowpart container is empty, the tree leak memory,
even when empty. This patch will remove the high part from the
HighLowContainer when the lowpart container is empty after
a removeLong operation.

Also, this highpart leak can lead to incorrect results when
checking for logical equality.