Fix Range Add Bug (#445)
This patch fixes a bug where Roaring64Bitmap will accept an invalid
range without failing and produce an incorrect bitmap. Since the range
is [inclusive, exclusive) adjusting the right end point by subtracting
one from it can create a bigger range. For example, bitmap.add(0, 0)
is an invalid range, but it will produce a bitmap with all positions
set to 1.